+++ CHANGES.md
@@ -11,7 +11,9 @@
 * No longer require new implementations to list populator modules in `implementations.__init__.py`
 * Implementation discovery is now done by keeping track of concrete implementations of the STACpopulatorBase class
 * Add ability to parse vertical data from CF metadata extracted from NCML files
-* Convert longitude values in range 0-360 degrees to -180-180 degrees to comply with WGS84
+* Convert coordinates using any CRS to to comply with WGS84 when building STAC bbox and geometries
+* Add option to provide a fallback CRS to use when one is not specified in the metadata
+* Add option to provide a CRS that will override the one specified in the metadata
 
 ## [0.12.0](https://github.com/crim-ca/stac-populator/tree/0.12.0) (2025-11-20)